third_party/blink/web_tests/animations/svg-attribute-interpolation/svg-d-interpolation-expected.txt

This is a testharness.js-based test.
[FAIL] SMIL: Interpolate attribute <d> from replace [m 10 0 h 1] to replace [m 20 0 v 2] at (0) is [m 10 0 h 1]
  assert_equals: expected "path ( \\" M 10 0 H 11 \\" ) " but got "path ( \\" M 10 0 \\" ) "
[FAIL] SMIL: Interpolate attribute <d> from replace [m 10 0 h 1] to replace [m 20 0 v 2] at (0.2) is [m 10 0 h 1]
  assert_equals: expected "path ( \\" M 10 0 H 11 \\" ) " but got "path ( \\" M 12 0 \\" ) "
[FAIL] SMIL: Interpolate attribute <d> from replace [m 10 0 h 1] to replace [m 20 0 v 2] at (0.6) is [m 20 0 v 2]
  assert_equals: expected "path ( \\" M 20 0 V 2 \\" ) " but got "path ( \\" M 16 0 \\" ) "
[FAIL] SMIL: Interpolate attribute <d> from replace [m 10 0 h 1] to replace [m 20 0 v 2] at (1) is [m 20 0 v 2]
  assert_equals: expected "path ( \\" M 20 0 V 2 \\" ) " but got "path ( \\" M 20 0 \\" ) "
Harness: the test ran to completion.
